don't query sources if using nginx
nginx processes shouldn't be halted, either by a curl http request or a
sleep when waiting for a metatile render, or a long mapserver render. We
thus return a 404 in libmapcache when a tile is not found, instead of
handling the metatile rendering ourselves. Nginx can be configured to
forward these kind of requests to a fastcgi or apache mapcache instance,
and the tiles that are created will be available for the next request.