command line option to switch between threads and processes
This commit enables the choice between using threads or (forked)
processes when using parallel seeding. The default is to use a single
thread. Support for multiple processes is disabled on windows platforms
due to the lack of ipc message queues.
Setting the size of the ipc message queue (msg_qbytes) might be
unsupported on osx, needs checking.