Allow S3 credentials and generic REST headers to be read at runtime
Allows security credentials to be read from an external file which
prevents them from having to be added to the main configuration file.
The use-case for this is an external script updating the credentials
file with temporary credentials obtained by some method.

The structure of the file for S3 is very simple:
- it may contain 2 or 3 lines, make sure to not add any trailing or
  leading whitespaces or newlines
- the first line is mandatory and is the AWS_ACCESS_KEY_ID
- the second line is mandatory and is the AWS_SECRET_ACCESS_KEY
- the third line is optional and may contain a security token (for
  temporary credentials)

The generic structure for all rest caches is a file referenced by
the <header_file> tag at the same level as <headers>. The file should
contain a list of headers to add, one per line, in the form:
header_name:header_value