source_gdal: make use of source dataset overviews.
For GDAL >= 2.0, this uses the capability of the VRTWarped mechanism
to expose overviews is the warped source dataset has overviews itself.
For GDAL < 2.0, a workaround is implemented to create a regular VRT
around the appropriate overview level of the source dataset.

Should work with GDAL >= 1.10 (and probably older),
although GDAL 2.0 or above is still recommended for speed and accuracy
in resampling methods.

Add support for resampleAlg and srcOvrLevel options in <gdalparams>

Accepted values for resampleAlg are: Near, Bilinear, Cubic, CubicSpline,
Lanczos and (GDAL >= 2.0) Average. Default is Bilinear.

srcOvrLevel is only supported for GDAL >= 2.0 and accepted values
are: AUTO (default), NONE, AUTO-xx (AUTO-1 means the first overview "above"
the AUTO one, ie an overview whose dimensions are larger), or xx
(xx=0 means first overview level).

e.g :
    <source name="global-tif" type="gdal">
      <data>world.tif</data>
      <gdalparams>
        <resampleAlg>Bilinear</resampleAlg>
        <srcOvrLevel>AUTO-1</srcOvrLevel>
      </gdalparams>
    </source>